Refactor checks to fit common/core/bridge pattern
