+++ typescript/optics-deploy/src/checks.ts
@@ -1,21 +1,8 @@
-import { CoreDeploy } from './core/CoreDeploy';
-import { checkCoreDeploy } from './core/checks';
-import { UpgradeBeacon, UpgradeBeaconController } from '@optics-xyz/ts-interface/dist/optics-core';
-
-export async function checkCoreDeploys(
-  coreDeploys: CoreDeploy[],
-  governorDomain: number,
-  invariantViolationHandler: InvariantViolationHandler
-) {
-  const checkDeploy = async (deploy: CoreDeploy) => {
-    const remoteDomains = coreDeploys.filter(_ => _.chain.domain !== deploy.chain.domain).map(_ => _.chain.domain)
-
-    console.info(`Checking core deploy on ${deploy.chain.name}`)
-    return checkCoreDeploy(deploy, remoteDomains, governorDomain, invariantViolationHandler)
-  }
-
-  await Promise.all(coreDeploys.map(checkDeploy))
-}
+import { expect } from 'chai';
+import { Contract, ethers } from 'ethers';
+import { Deploy } from './deploy';
+import { ProxyNames, BeaconProxy } from './proxyUtils';
+import { UpgradeBeaconController } from '@optics-xyz/ts-interface/dist/optics-core';
 
 export enum InvariantViolationType {
   UpgradeBeacon
@@ -23,21 +10,22 @@ export enum InvariantViolationType {
 
 interface UpgradeBeaconInvariantViolation {
   domain: number
+  name: ProxyNames
   upgradeBeaconController: UpgradeBeaconController,
   type: InvariantViolationType.UpgradeBeacon,
-  beacon: UpgradeBeacon,
+  beaconProxy: BeaconProxy<ethers.Contract>,
   expectedImplementationAddress: string
   actualImplementationAddress: string
 }
 
-type InvariantViolation = UpgradeBeaconInvariantViolation
+export type InvariantViolation = UpgradeBeaconInvariantViolation
 
 export type InvariantViolationHandler = (violation: InvariantViolation) => void
 
 export const assertInvariantViolation = (violation: InvariantViolation) => {
   switch (violation.type) {
     case InvariantViolationType.UpgradeBeacon:
-      throw new Error(`Expected BeaconProxy at address at ${violation.beacon.address} to point to implementation at ${violation.expectedImplementationAddress}, found ${violation.actualImplementationAddress}`)
+      throw new Error(`Expected UpgradeBeacon at address at ${violation.beaconProxy.beacon.address} to point to implementation at ${violation.expectedImplementationAddress}, found ${violation.actualImplementationAddress}`)
       break;
     default:
       break;
@@ -59,7 +47,57 @@ export class InvariantViolationCollector {
       m.actualImplementationAddress === v.actualImplementationAddress &&
       m.expectedImplementationAddress === v.expectedImplementationAddress
     )
-    if (duplicateIndex !== -1)
+    if (duplicateIndex === -1)
       this.violations.push(v);
   }
 }
+
+export function checkVerificationInput(
+  deploy: Deploy<any>,
+  name: string,
+  addr: string,
+) {
+  const match = deploy.verificationInput.find(
+    (contract) => contract.name == name && contract.address === addr
+  )
+  expect(match).to.not.be.undefined;
+}
+
+export function assertBeaconProxy(beaconProxy: BeaconProxy<Contract>) {
+  expect(beaconProxy.beacon).to.not.be.undefined;
+  expect(beaconProxy.proxy).to.not.be.undefined;
+  expect(beaconProxy.implementation).to.not.be.undefined;
+}
+
+export async function checkBeaconProxyImplementation(
+  domain: number,
+  name: ProxyNames,
+  upgradeBeaconController: UpgradeBeaconController,
+  beaconProxy: BeaconProxy<Contract>,
+  invariantViolationHandler: InvariantViolationHandler,
+) {
+  assertBeaconProxy(beaconProxy)
+
+  // Assert that the implementation is actually set
+  const provider = beaconProxy.beacon.provider;
+  const storageValue = await provider.getStorageAt(
+    beaconProxy.beacon.address,
+    0,
+  );
+  const actualImplementationAddress = ethers.utils.getAddress(
+    storageValue.slice(26),
+  );
+
+  if (actualImplementationAddress != beaconProxy.implementation.address) {
+    invariantViolationHandler({
+      type: InvariantViolationType.UpgradeBeacon,
+      name,
+      domain,
+      upgradeBeaconController,
+      beaconProxy,
+      actualImplementationAddress,
+      expectedImplementationAddress: beaconProxy.implementation.address,
+    });
+  }
+}
+
